+++ Mythos/Mythos/View/GamePlay/View/MaybeGameView.swift
@@ -10,19 +10,14 @@ import CoreHaptics
 
 struct MaybeGameView: View {
     @EnvironmentObject var websocket: WebSocket
-
     @ObservedObject var cardVM = CardViewModel()
 
-//    @State var cardSelected: Card? = nil
-//    @State var isTapped: Bool = false
-//    @State var killTapped: Bool = false
+    @Environment(\.dismiss) private var dismiss
 
     @State var isPresentedGame: Bool
     @State var showAlertWinner: Bool = false
     @State var showAlertLost: Bool = false
     @State private var duoConditionalALert: Bool = false
-
-    @Environment(\.dismiss) private var dismiss
 
     var body: some View {
         ZStack {
@@ -31,11 +26,7 @@ struct MaybeGameView: View {
                 .ignoresSafeArea()
             VStack {
                 Spacer()
-                ForEach(Array(websocket.connectedPlayers.enumerated()), id: \.element.id) { index, player in
-                    if websocket.myPlayerReference == player {
-                            generatePlayerLayout(for: index, players: websocket.connectedPlayers)
-                    }
-                }
+                PlayerLayoutView(cardVM: cardVM)
             }
             .overlay {
                 VStack {
@@ -43,10 +34,7 @@ struct MaybeGameView: View {
                 }
             }
             .onChange(of: websocket.isGameOver) { newValue in
-                print("O jogador \(websocket.myPlayerReference.name) perdeu a partida")
-
                 showAlertLost = true
-                print("perdeu : \(showAlertLost) ganhou : \(showAlertWinner)")
                 isPresentedGame = false
 
                 UserDefaults.standard.set(showAlertLost, forKey: "lost")
@@ -56,10 +44,7 @@ struct MaybeGameView: View {
                 UserDefaults.standard.set(isPresentedGame, forKey: "isPresentedGame")
             }
             .onChange(of: websocket.winner) { newValue in
-                print("O jogador \(websocket.myPlayerReference.name) venceu a partida")
                 showAlertWinner = true
-                print("perdeu : \(showAlertLost) ganhou : \(showAlertWinner)")
-
                 UserDefaults.standard.set(showAlertWinner, forKey: "win")
             }
             .alert(isPresented: $duoConditionalALert) {
@@ -78,104 +63,23 @@ struct MaybeGameView: View {
         .overlay {
             //MARK: - Mais detalhes cartas da mão
             if cardVM.isTapped == true {
-                ZStack {
-                    Rectangle()
-                        .foregroundColor(.black)
-                        .opacity(0.5)
-                        .ignoresSafeArea()
-                        .onTapGesture {
-                            withAnimation {
-                                cardVM.isTapped = false
-                            }
+                CardHighlightsView(cardVM: cardVM, offSetValuex: 0, offSetValuey: -20, card: cardVM.cardSelected!)
+                    .onTapGesture {
+                        withAnimation {
+                            cardVM.isTapped = false
                         }
-                    CardFocusedView(card: cardVM.cardSelected!, isTapped: $cardVM.isTapped)
-                        .frame(width: 744/3.5, height: 1039/3.5)
-                        .offset(y: -20)
-                }
+                    }
             }
             //MARK: - Botão jogar cartas
             if websocket.myPlayerReference.isYourTurn {
-                Button {
-                    websocket.sendCard(with: cardVM.cardSelected)
-                    self.cardVM.isTapped = false
-                    self.cardVM.killTapped = false
-                } label: {
-                    Image("button_descarte")
-                        .frame(maxWidth: 40, maxHeight: 40)
-                        .scaledToFit()
-                }
-                .offset(x: 250, y: 110)
-                .disabled(!cardVM.isTapped)
+                SendCardButtonView(cardVM: cardVM)
             }
             //MARK: - Clique da Carta Cemitério
             if cardVM.killTapped == true {
-                ZStack {
-                    Rectangle()
-                        .foregroundColor(.black)
-                        .opacity(0.5)
-                        .ignoresSafeArea()
-                    CardFocusedView(card: websocket.cardsPlayed.last!, isTapped: $cardVM.killTapped)
-                        .frame(width: 744/3.5, height: 1039/3.5)
-                        .offset(x: 130, y:0)
-                }
+                CardHighlightsView(cardVM: cardVM, offSetValuex: 130, offSetValuey: 0, card: websocket.cardsPlayed.last!)
             }
         }
         .navigationBarBackButtonHidden(true)
-    }
-    
-    func generatePlayerLayout(for index: Int, players: [PlayerClient]) -> some View {
-
-        let firstPlayerIndex = (index) % players.count
-        let secondPlayerIndex = (index + 1) % players.count
-        let thirdPlayerIndex = (index + 2) % players.count
-        let lastPlayerIndex = (index + 3) % players.count
-
-        let firstPlayer = players[firstPlayerIndex]
-        let secondPlayer = players[secondPlayerIndex]
-        let thirdPlayer = players[thirdPlayerIndex]
-        let lastPlayer = players[lastPlayerIndex]
-
-        var viewPersonas: some View {
-            VStack {
-                PersonasView(cards: thirdPlayer.handCards,
-                             namePerson: thirdPlayer.name,
-                             lifePerson: (thirdPlayer.life <= 0) ? 0 : thirdPlayer.life,
-                             index: (players.count < 3) ? 5 : 2)
-
-                HStack {
-                    PersonasView(cards: secondPlayer.handCards,
-                                 namePerson: secondPlayer.name,
-                                 lifePerson: (secondPlayer.life <= 0) ? 0 : secondPlayer.life,
-                                 index: (players.count < 2) ? 5 : 1)
-                    Spacer()
-
-                    PersonasView(cards: lastPlayer.handCards, namePerson: lastPlayer.name,
-                                 lifePerson: (lastPlayer.life <= 0) ? 0 : lastPlayer.life,
-                                 index: (players.count < 4) ? 5 : 3)
-
-                }
-                .overlay {
-                    if (websocket.cardsPlayed.last != nil) {
-                        Button {
-                            withAnimation {
-                                cardVM.killTapped.toggle()
-                                cardVM.isTapped = false
-                            }
-                        } label: {
-                            KillDeckView(card: websocket.cardsPlayed.last!, killDecktapped: $cardVM.killTapped)
-                                .frame(width: 744/9, height: 1039/9)
-                                .offset(x: cardVM.killTapped ? 110 : 0, y:0)
-                                .opacity(cardVM.killTapped ? 0 : 1)
-                        }
-                    }
-                }
-                PersonasView(cards: firstPlayer.handCards,
-                             namePerson: firstPlayer.name,
-                             lifePerson: (firstPlayer.life <= 0) ? 0 : firstPlayer.life,
-                             index: 0)
-            }
-        }
-        return viewPersonas
     }
 }
 
@@ -186,30 +90,4 @@ struct MaybeGameView: View {
 //    }
 //}
 
-struct GaugeProgressStyle: ProgressViewStyle {
-    var strokeColor = Color.green.gradient
-    var strokeWidth = 8.0
-
-    func makeBody(configuration: Configuration) -> some View {
-        let fractionCompleted = configuration.fractionCompleted ?? 0
-        let number = (configuration.fractionCompleted ?? 0) * 30
-
-        return ZStack {
-            Circle()
-                .trim(from: 0, to: fractionCompleted)
-                .stroke(strokeColor, style: StrokeStyle(lineWidth: strokeWidth, lineCap: .round))
-                .rotationEffect(.degrees(-90))
-                .overlay {
-                    VStack {
-                        Text("Sua vida:")
-                            .font(Font.title3.bold())
-                            .foregroundColor(.white)
-                        Text(Int(number).description)
-                            .font(Font.title3.bold())
-                            .foregroundColor(.white)
-                    }
-                }
-        }
-    }
-}
